+++ app/controllers/api/v1/battlegroups_controller.rb
@@ -8,7 +8,8 @@ module Api
 
       def index
         authorize current_player
-        @pagy, @battlegroups = pagy(Battlegroup.includes(:qc_attacks).where(player_id: current_player.id))
+
+        @pagy, @battlegroups = pagy(Battlegroup.where(player_id: current_player.id))
 
         return unless stale? @battlegroups