+++ app/javascript/lca/ducks/entities/_lib.ts
@@ -43,6 +43,7 @@ type crudActions =
   | 'REGEN_CODE'
   | 'ADD_THING'
   | 'REMOVE_THING'
+  | 'FETCH_FOR_CHRONICLE'
 
 export const API = 'lca-api'